Update 25/01 - library

+++ app/src/main/java/com/example/doantotnghiep/data/model/SensorData.kt
@@ -1,7 +1,7 @@
 package com.example.doantotnghiep.data.model
 
 data class SensorData(
-    val timestamp: Int?,
+    val timestamp: Long?,
     val distanceRaw: Int?,
     val temp: Double?,
     val humid: Double?,
